+++ bdtsim/cli/list_protocols.py
@@ -22,6 +22,8 @@ from .command_manager import SubCommand
 
 
 class ListProtocolsSubCommand(SubCommand):
+    help = 'print a list of supported protocols to be simulated'
+
     def __call__(self, args: argparse.Namespace) -> None:
         for name in ProtocolManager.protocols.keys():
             print(name)