+++ bdtsim/funds_diff_collection.py
@@ -20,47 +20,74 @@ from typing import Any, Dict, Optional
 from .account import Account
 
 
-class FundsDiffCollection(object):
-    def __init__(self, initial: Optional[Dict[Account, int]] = None) -> None:
-        if initial is not None:
-            self._dict = initial
+class FundsDiffCollection(Dict[Account, int]):
+    def __init__(self, init: Optional[Dict[Account, int]] = None) -> None:
+        if init is None:
+            super().__init__()
         else:
-            self._dict = {}
+            super().__init__(init)
 
-    def get(self, account: Account) -> int:
-        diff = self._dict.get(account)
-        if diff is None:
-            return 0
-        else:
-            return diff
+    def get(self, account: Account, default: int = 0) -> int:  # type: ignore
+        return super().get(account, default)
 
     def __repr__(self) -> str:
         return '<%s.%s %s>' % (
             __name__,
             FundsDiffCollection.__name__,
-            ', '.join(['%s: %d' % (account.name, diff) for account, diff in self._dict.items()])
+            ', '.join(['%s: %d' % (account.name, diff) for account, diff in self.items()])
         )
 
     def __iadd__(self, other: Any) -> 'FundsDiffCollection':
         if isinstance(other, FundsDiffCollection):
-            for account, other_amount in other._dict.items():
-                current_amount = self._dict.get(account)
+            for account, other_amount in other.items():
+                current_amount = self.get(account)
                 if current_amount is None:
-                    self._dict.update({account: other_amount})
+                    self.update({account: other_amount})
                 else:
-                    self._dict.update({account: current_amount + other_amount})
+                    self.update({account: current_amount + other_amount})
             return self
+        else:
+            return NotImplemented
+
+    def __add__(self, other: Any) -> 'FundsDiffCollection':
+        if isinstance(other, FundsDiffCollection):
+            result = FundsDiffCollection()
+            result += self
+            result += other
+            return result
         else:
             return NotImplemented
 
     def __isub__(self, other: Any) -> 'FundsDiffCollection':
         if isinstance(other, FundsDiffCollection):
-            for account, other_amount in other._dict.items():
-                current_amount = self._dict.get(account)
+            for account, other_amount in other.items():
+                current_amount = self.get(account)
                 if current_amount is None:
-                    self._dict.update({account: other_amount})
+                    self.update({account: -other_amount})
                 else:
-                    self._dict.update({account: current_amount - other_amount})
+                    self.update({account: current_amount - other_amount})
             return self
         else:
             return NotImplemented
+
+    def __sub__(self, other: Any) -> 'FundsDiffCollection':
+        if isinstance(other, FundsDiffCollection):
+            result = FundsDiffCollection()
+            result += self
+            result -= other
+            return result
+        else:
+            return NotImplemented
+
+    def __eq__(self, other: Any) -> bool:
+        if isinstance(other, FundsDiffCollection):
+            for a, b in (self, other), (other, self):
+                for account in a.keys():
+                    if a.get(account) != b.get(account):
+                        return False
+            return True
+        else:
+            return NotImplemented
+
+    def __ne__(self, other: Any) -> bool:
+        return not self.__eq__(other)