+++ bdtsim/output/simulation_result.py
@@ -17,20 +17,20 @@
 
 import copy
 import itertools
-from typing import Any, Dict, List, Optional
+from typing import Any, Callable, Dict, List, Optional, Tuple
 
-from bdtsim.account import Account
+from bdtsim.account import Account, seller, buyer
 from bdtsim.funds_diff_collection import FundsDiffCollection
 from bdtsim.protocol_path import Decision
 
 
 class TransactionLogEntry(object):
     def __init__(self, account: Account, tx_dict: Dict[str, Any], tx_receipt: Dict[str, Any],
-                 funds_diff: FundsDiffCollection) -> None:
+                 funds_diff_collection: FundsDiffCollection) -> None:
         self.account = account
         self.tx_dict = tx_dict
         self.tx_receipt = tx_receipt
-        self.funds_diff = funds_diff
+        self.funds_diff_collection = funds_diff_collection
 
 
 class TransactionLogList(object):
@@ -53,6 +53,7 @@ class TransactionLogList(object):
         def __init__(self, tx_list: 'TransactionLogList') -> None:
             self.entries: Dict[Account, TransactionLogList.Aggregation.Entry] = {}
             for tx in tx_list.tx_log_list:
+                # work on log line
                 entry = self.entries.get(tx.account)
                 if entry is None:
                     self.entries.update({
@@ -60,13 +61,26 @@ class TransactionLogList(object):
                             tx.account,
                             int(tx.tx_receipt['gasUsed']),
                             1,
-                            tx.funds_diff.get(tx.account)
+                            0
                         )
                     })
                 else:
                     entry.tx_fees += int(tx.tx_receipt['gasUsed'])
                     entry.tx_count += 1
-                    entry.funds_diff += tx.funds_diff.get(tx.account)
+                # work on all fund diffs
+                for account, funds_diff in tx.funds_diff_collection.items():
+                    funds_diff_entry = self.entries.get(account)
+                    if funds_diff_entry is None:
+                        self.entries.update({
+                            account: TransactionLogList.Aggregation.Entry(
+                                account,
+                                0,
+                                0,
+                                funds_diff
+                            )
+                        })
+                    else:
+                        funds_diff_entry.funds_diff += funds_diff
 
     def __init__(self) -> None:
         self.tx_log_list: List[TransactionLogEntry] = []
@@ -242,8 +256,60 @@ class ResultNode(object):
         return aggregation_summary
 
 
+class AggregationAttributeHelper(object):
+    def __init__(self, account: Account, attribute: str) -> None:
+        self._account = account
+        self._attribute = attribute
+
+    def __call__(self, aggregation: TransactionLogCollection.Aggregation) -> Any:
+        return getattr(aggregation.entries.get(self._account), self._attribute)
+
+
 class SimulationResult(object):
     def __init__(self) -> None:
         self.preparation_transactions = TransactionLogList()
         self.execution_result_root = ResultNode()
         self.cleanup_transactions = TransactionLogList()
+
+    def get_important_execution_results(self) -> List[Tuple[str, TransactionLogCollection.Aggregation]]:
+        all_honest_results = []
+        seller_honest_results = []
+        buyer_honest_results = []
+
+        for node in self.execution_result_root.final_nodes:
+            if node.all_accounts_completely_honest():
+                all_honest_results.append(node.aggregation_summary)
+            elif node.account_completely_honest(seller):
+                seller_honest_results.append(node.aggregation_summary)
+            elif node.account_completely_honest(buyer):
+                buyer_honest_results.append(node.aggregation_summary)
+
+        important_results = []
+        for honesty_str, results in [
+            ('All Parties Honest', all_honest_results),
+            ('Honest Seller, Cheating Buyer', seller_honest_results),
+            ('Honest Buyer, Cheating Seller', buyer_honest_results)
+        ]:
+            for account in seller, buyer:
+                tx_fees_max_result = self._apply_aggr_func(max, results, account, 'tx_fees_max')
+                if tx_fees_max_result is not None:
+                    important_results.append(('%s - %s max fees' % (honesty_str, account.name), tx_fees_max_result))
+
+            profit_max_result = self._apply_aggr_func(max, results, seller, 'value_diff_max')
+            if profit_max_result is not None:
+                important_results.append(('%s - Seller max profit' % honesty_str, profit_max_result))
+
+            expenses_max_result = self._apply_aggr_func(min, results, buyer, 'value_diff_min')
+            if expenses_max_result is not None:
+                important_results.append(('%s - Buyer max expenses' % honesty_str, expenses_max_result))
+
+        return important_results
+
+    @staticmethod
+    def _apply_aggr_func(aggr_func: Callable[..., TransactionLogCollection.Aggregation],
+                         results: List[TransactionLogCollection.Aggregation], account: Account,
+                         attribute: str) -> Optional[TransactionLogCollection.Aggregation]:
+        results_filtered = list(filter(lambda x: x.entries.get(account) is not None, results))
+        if len(results_filtered) == 0:
+            return None
+        return aggr_func(results_filtered, key=AggregationAttributeHelper(account, attribute))