+++ bdtsim/protocol/fairswap/fairswap.py
@@ -117,7 +117,7 @@ class FairSwap(Protocol):
                                          ))
 
         plain_data = data_provider.file_pointer.read()
-        plain_merkle_tree = merkle.from_bytes(plain_data)
+        plain_merkle_tree = merkle.from_bytes(plain_data, self._slices_count)
         key = self.generate_bytes(32, 1337)
 
         # === 1a: Seller: encrypt file for transmission
@@ -127,18 +127,31 @@ class FairSwap(Protocol):
         if encryption_decision == 'expected':
             encrypted_merkle_tree = encoding.encode(plain_merkle_tree, key)
         elif encryption_decision == 'completely different':
-            encrypted_merkle_tree = encoding.encode(merkle.from_bytes(self.generate_bytes(data_provider.data_size)),
-                                                    key)
+            encrypted_merkle_tree = encoding.encode(
+                merkle.from_bytes(self.generate_bytes(data_provider.data_size), self._slices_count),
+                key
+            )
         elif encryption_decision == 'leaf forgery' or encryption_decision == 'hash forgery':
-            data_leaves = [leaf.data for leaf in plain_merkle_tree.leaves]
-            data_digests = plain_merkle_tree.digests_pack
-            data_leaves[0] = b'\x00' * len(data_leaves[0])
+            # extract correct plain data
+            plain_leaves_data = [leaf.data for leaf in plain_merkle_tree.leaves]
+            plain_digests = plain_merkle_tree.digests_pack
+
+            # forge first leaf
+            plain_leaves_data[0] = b'\x00' * len(plain_leaves_data[0])
 
             if encryption_decision == 'hash forgery':
-                data_digests[0] = Web3.solidityKeccak(['bytes32', 'bytes32'], [data_leaves[0], data_leaves[1]])
+                plain_digests[0] = merkle.MerkleTreeNode(
+                    merkle.MerkleTreeLeaf(plain_leaves_data[0]),
+                    merkle.MerkleTreeLeaf(plain_leaves_data[1])
+                ).digest
 
-            encrypted_merkle_tree = merkle.from_leaves([merkle.MerkleTreeLeaf(x) for x in data_leaves]
-                                                       + [merkle.MerkleTreeHashLeaf(x) for x in data_digests]
+            encrypted_leaves_data = [encoding.crypt(data, i, key) for i, data in enumerate(plain_leaves_data)]
+            encrypted_digests = [
+                encoding.crypt(data, 2 * len(plain_leaves_data) + i, key) for i, data in enumerate(plain_digests)
+            ]
+
+            encrypted_merkle_tree = merkle.from_leaves([merkle.MerkleTreeLeaf(x) for x in encrypted_leaves_data]
+                                                       + [merkle.MerkleTreeHashLeaf(x) for x in encrypted_digests]
                                                        + [merkle.MerkleTreeHashLeaf(encoding.B032)])
         else:
             raise NotImplementedError()
@@ -204,37 +217,55 @@ class FairSwap(Protocol):
             raise NotImplementedError()
 
         # === 4: Buyer: Send Ok/Complain/Leave
-        try:
-            decrypted_merkle_tree = encoding.decode(encrypted_merkle_tree, transfer_key)
-            if decrypted_merkle_tree.digest == plain_merkle_tree.digest:
-                if protocol_path.decide(buyer, 'Confirm', variants=['yes', 'leave']) == 'yes':
+        root_hash_enc = encrypted_merkle_tree.leaves[-2].data
+        if encoding.crypt(root_hash_enc, 3 * self._slices_count - 2, transfer_key) != plain_merkle_tree.digest:
+            if protocol_path.decide(buyer, 'complain about root', ['yes']) == 'yes':
+                logger.debug('Buyer: Complaining about incorrect file root hash')
+                root_hash_leaf = encrypted_merkle_tree.leaves[-2]
+                proof = encrypted_merkle_tree.get_proof(root_hash_leaf)
+                environment.send_contract_transaction(buyer, 'complainAboutRoot', root_hash_leaf.data, proof)
+                return
+        else:
+            decrypted_merkle_tree, errors = encoding.decode(encrypted_merkle_tree, transfer_key)
+            if len(errors) == 0:
+                if protocol_path.decide(buyer, 'Confirm', variants=['yes', 'leave'],
+                                        honest_variants=['yes', 'leave']) == 'yes':
                     environment.send_contract_transaction(buyer, 'noComplain')
                     return
-            else:
-                if protocol_path.decide(buyer, 'complain about root', ['yes']) == 'yes':
-                    logger.debug('Buyer: Complaining about incorrect file root hash')
-                    root_hash_leaf = encrypted_merkle_tree.leaves[-2]
-                    proof = encrypted_merkle_tree.get_proof(root_hash_leaf)
-                    environment.send_contract_transaction(buyer, 'complainAboutRoot', root_hash_leaf.data, proof)
+            elif isinstance(errors[-1], encoding.LeafDigestMismatchError):
+                if protocol_path.decide(buyer, 'Complain about Leaf', ['yes']) == 'yes':
+                    error: encoding.NodeDigestMismatchError = errors[-1]
+                    environment.send_contract_transaction(
+                        buyer,
+                        'complainAboutLeaf',
+                        error.index_out,
+                        error.index_in,
+                        error.out.data,
+                        error.in1.data_as_list(),
+                        error.in2.data_as_list(),
+                        encrypted_merkle_tree.get_proof(error.out),
+                        encrypted_merkle_tree.get_proof(error.in1)
+                    )
                     return
-        except encoding.DigestMismatchError as error:
-            complain_subject = 'Leaf' if error.index_in < self._slices_count else 'Node'
-            if protocol_path.decide(buyer, 'Complain about %s' % complain_subject, ['yes']) == 'yes':
-                environment.send_contract_transaction(
-                    buyer,
-                    'complainAbout%s' % complain_subject,
-                    error.index_out,
-                    error.index_in,
-                    error.out.data,
-                    error.in1.data_as_list(),
-                    error.in2.data_as_list(),
-                    encrypted_merkle_tree.get_proof(error.out),
-                    encrypted_merkle_tree.get_proof(error.in1)
-                )
-                return
+            else:
+                if protocol_path.decide(buyer, 'Complain about Node', ['yes']) == 'yes':
+                    error = errors[-1]
+                    environment.send_contract_transaction(
+                        buyer,
+                        'complainAboutNode',
+                        error.index_out,
+                        error.index_in,
+                        error.out.data,
+                        error.in1.data,
+                        error.in2.data,
+                        encrypted_merkle_tree.get_proof(error.out),
+                        encrypted_merkle_tree.get_proof(error.in1)
+                    )
+                    return
 
         # === 5: Seller: Finalize (when Buyer leaves in 4)
-        if protocol_path.decide(seller, 'Request Payout', variants=['yes', 'no']) == 'yes':
+        if protocol_path.decide(seller, 'Request Payout', variants=['yes', 'no'],
+                                honest_variants=['yes', 'no']) == 'yes':
             environment.wait(self._timeout + 1)
             environment.send_contract_transaction(seller, 'refund')